requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
imgcompare==2.0.1
jsonfield==3.1.0
kombu==5.0.2
Markdown==3.3.3
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
vk==2.0.2
vk-api==11.9.1
wcwidth==0.2.5
whitenoise==5.2.0
